Comma handling scenarios didn't cater for non-Name types.
Moved tests to their own class.
